+++ settings.gradle
@@ -18,9 +18,9 @@ include 'thunx-dependencies'
 include 'thunx-model'
 include 'thunx-pdp'
 include 'thunx-pdp-opa'
-include 'thunx-security'
 include 'thunx-spring-api'
 include 'thunx-spring-gateway'
+include 'thunx-spring-security'
 include 'spring-data-querydsl-predicate-injector'
 include 'thunx-autoconfigure'
 include 'thunx-api-spring-boot-starter'
