+++ thunx-bom/build.gradle
@@ -13,9 +13,9 @@ dependencies {
         api project(':thunx-pdp')
         api project(':thunx-pdp-opa')
         api project(':thunx-predicates-querydsl')
-        api project(':thunx-security')
         api project(':thunx-spring-api')
         api project(':thunx-spring-gateway')
+        api project(':thunx-spring-security')
         api project(':thunx-visitor-reducer')
 
         api project(':spring-data-querydsl-predicate-injector')
