+++ thunx-spring-security/src/main/java/com/contentgrid/thunx/spring/security/AbacContextAuthority.java
@@ -1,4 +1,4 @@
-package com.contentgrid.thunx.security;
+package com.contentgrid.thunx.spring.security;
 
 import com.contentgrid.thunx.predicates.model.ThunkExpression;
 import lombok.Data;
