+++ thunx-spring-security/src/test/java/com/contentgrid/thunx/spring/security/AbacJwtGrantedAuthoritiesConverterTest.java
@@ -1,4 +1,4 @@
-package com.contentgrid.thunx.security;
+package com.contentgrid.thunx.spring.security;
 
 import static org.assertj.core.api.Assertions.assertThat;
